📦 Moved EmployeeRow functionality out
